migrated FindVolumeTemplate Moved to check for upgrade and install
